mysite/blog/requirements.txt
asgiref==3.2.3
autopep8==1.5
django-debug-toolbar==2.2
django-markdown==0.8.4
django-taggit==1.2.0
Markdown==3.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
